logistic upd

+++ cycletest.sh
@@ -1,20 +1,37 @@
+#check input var
+if [ "$#" -ne 1 ]
+then
+	echo "[ERROR]Invalid number of arguments. Needed one (number of tests)"
+	exit
+fi
+
+#pre
+unt=$1
+((unt++))
 rm fails
 touch fails
 echo "If the file is not empty, HAHAHHA LOH" > fails
-for ((i = 1; i < 2; i++))
+count=0
+
+make -C ./resources/ test
+
+#main alg
+for ((i = 1; i < unt; i++))
 do
-	echo $i
+	printf $i"|"
 
 	touch differ
 
-	key=`cat -n key`
-	iv=`cat -n iv`
-	make test
-	./tester
-	sh test.sh $key $iv #> silence
+	#one step of test
+	key=`cat -n ./resources/files/key`
+	iv=`cat -n ./resources/files/iv`
+	./resources/tester
+	sh ./resources/bash_scripts/test.sh $key $iv 
 	err=`wc -l < differ`
+	#if there is a difference
 	if ! [ "$err" -eq "0" ]
 	then
+		((count++))
 		echo " ===================== " >> fails
 		printf ">>> TEXT \"" >> fails
 		cat plaintext >> fails
@@ -29,7 +46,19 @@ do
 		cat differ >>fails
 		echo "" >> fails
 	fi
-
 	rm differ
-
 done
+
+# result print
+if ! [ "$unt" -gt 1 ]
+then
+	exit
+fi
+
+echo ""
+if ! [ "$count" -eq "0" ]
+then
+	echo '\033[0;31m'$count" FAILS!"'\033[0m'
+else
+	echo '\033[0;32m'"OK :)"'\033[0m'
+fi
